+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_heating_curve_offset()` to heating circuit
 - Added endpoint `get_heating_curve_slope()` to heating circuit
 - Added endpoint `set_heating_curve_slope()` to heating circuit
+- Added endpoint `get_use_heating_curve()` to heating circuit
+- Added endpoint `set_use_heating_curve()` to heating circuit
 
 ## [2.8.1] - 2026-01-28